sso: move singleflight package into internal/pkg
